requirements.txt
google-generativeai
python-dotenv
streamlit
pymongo[srv]
tavily-python